Increase default move speed & allow rotation speed to be adjusted
Both the default movement speed and the (unadjustable) rotational
speed are way too slow. Not sure if this is intended.

I added a logarithmic formula for the rotational speed based on
movement speed. Not fully sure if this formula is the best but it
seems to work decently intuitively to me.